% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{segm_connected_components}
\alias{segm_connected_components}
\title{Connected components segmentation}
\usage{
segm_connected_components(
  image,
  otb,
  expr = "distance > 10",
  mode = "vector",
  vector_neighbor = FALSE,
  vector_stitch = TRUE,
  vector_minsize = 1L,
  vector_simplify = 0.1,
  vector_tilesize = 1024L,
  mask = NULL
)
}
\arguments{
\item{image}{path or \code{SpatRaster}}

\item{otb}{output of link2GI::linkOTB()}

\item{expr}{user defined connection condition, written as a mathematical expression.
Available variables are 'p(i)b(i)', 'intensity_p(i)', and 'distance'. Substitute
(i) by the desired value (e.g. 'intensity_p2 Z 0.5')}

\item{mode}{processing mode, either 'vector' or 'raster'. See details}

\item{vector_neighbor}{logical. If FALSE (the default) a 4-neighborhood connectivity
is activated. If TRUE, a 8-neighborhood connectivity is used}

\item{vector_stitch}{logical. If TRUE (the default), scans polygons on each side
of tiles and stitch polygons which connect by more than one pixel}

\item{vector_minsize}{integer. Objects whose size in pixels is below the minimum
object size will be ignored during vectorization}

\item{vector_simplify}{simplify polygons according to a given tolerance (in pixel).
This option allows reducing the size of the output file or database.}

\item{vector_tilesize}{integer. User defined tiles size for tile-based segmentation.
Optimal tile size is selected according to available RAM if NULL}

\item{mask}{an optional raster used for masking the segmentation. Only pixels
whose mask is strictly positive will be segmented}
}
\value{
\code{sf} or \code{SpatRaster}
}
\description{
Applies the connected components segmentation algorithm to an image file
or a SpatRaster
}
\details{
The connected components segmentation algorithm groups neighboring pixels with
similar properties into distinct labeled regions. This method works by identifying
sets of adjacent pixels that share the same value or meet a prefined similarity
criterion. Steps:
\enumerate{
\item Initialization: Each pixel is analyzed to determine if it shares a connection
(4- or 8-connectivity) with its neighbors based on intensity, color,
or other attributes.
\item Grouping: Pixels connected by similarity form a region, and each region is
assigned a unique label.
\item Label Propagation: The algorithm iterates over the image, assigning the same
label to all connected pixels within a region.
\item Output: A segmented image is generated where each region has a unique identifier.
}

The processing mode 'vector' will output a vector file, and process the input
image piecewise. This allows performing segmentation of very large images. IN
contrast, 'raster' mode will output a labeled raster, and it cannot handle
large data. If mode is 'raster', all the 'vector_*' arguments are ignored.
}
